Block namespace search (#117)
It is possible to search for pages that are not articles
by using a namespce preffix in the search terms.
For instance: Help:Style, User_talk:SomeUserName,
or even Special:AbuseFilter

None of those results are relevant for this application.

This commit blocks those by replacing all colons with spaces.

It is possible for an article title to contain a colon but
very few do. And those who do generally have a redirect
from a similar title without it. I've included the "redirects"
API option to resolve redirects and mitigate the impact of
ommiting the colon.